backend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.2
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cors==1.0.1
decorator==5.1.1
distlib==0.3.6
Django==4.2.1
django-apscheduler==0.6.2
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-model-utils==4.3.1
django-notifications-hq==1.8.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.10.7
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.3.3
future==0.18.3
geopy==2.3.0
gevent==23.9.1
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
googleapis-common-protos==1.62.0
greenlet==3.0.0rc3
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonfield==3.1.0
kombu==5.3.1
MarkupSafe==2.1.3
multidict==6.0.4
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psycopg2==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.1
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
ratelim==0.1.6
redis==4.6.0
requests==2.31.0
requests-file==1.5.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.21
sqlparse==0.4.3
swapper==1.3.0
tldextract==3.6.0
typing_extensions==4.8.0
tzdata==2023.2
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
virtualenv==20.21.0
wcwidth==0.2.6
Werkzeug==2.3.7
whitenoise==6.5.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
